Generating a Python Library
At this point the tutorials and docs get you `nx generate ...` to create new projects. There are ~20 CLI parameters for the python poetry generator, and it doesn't seem useful to learn about them or juggle between the CLI and the docs. Instead, I've used the VS Code extension by just clicking "generate" in the sidebar, which has then given me a nice form listing all the options (which I guess is generated from a schema provided by the author).

Here's the command it ended up generating for me:

```
@nxlv/python:poetry-project --name=pythonlib --projectType=library --directory=python/libs/pythonlib --codeCoverage=false --codeCoverageHtmlReport=false --codeCoverageXmlReport=false --linter=none --projectNameAndRootFormat=as-provided --pyenvPythonVersion=3.11.2 --pyprojectPythonDependency=">=3.9,<3.12" --unitTestHtmlReport=false --unitTestJUnitReport=false --unitTestRunner=none --no-interactive --dry-run
```

Notable things:

- I've disabled lints, tests and code coverage to keep this minimal. I'll want to learn how to add those myself later.
- I bumped the Python version because I don't have the default version installed.
- I've set the project directory to include a `python/libs` prefix because I want to keep different languages separate in future.